+++ model/hangmanGuessedLetter.go
@@ -2,7 +2,7 @@ package model
 
 type HangmanGuessedLetter struct {
 	Base
-	Id        uint   `gorm:"column:id; primary_key"`
-	HangmanId uint   `gorm:"column:hangman_id"`
+	Id        int64   `gorm:"column:id; primary_key"`
+	HangmanId int64   `gorm:"column:hangman_id"`
 	Letter    string `gorm:"column:letter"`
 }
